Refactor schema imports in user API files for improved clarity. This change standardizes import comments and enhances code readability across fetch-user-expenses.ts, fetch-users-by-ids.ts, and fetch-users.ts.
